requirements.txt
akshare>=1.12.0
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
thsdk>=0.0.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0

langchain>=0.2.14
langchain-openai>=0.1.20
pydantic>=2.0.0
